backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
pydantic-settings>=2.0.0
python-multipart>=0.0.12
numpy>=1.26.0
pandas>=2.2.0
prophet>=1.1.5
holidays>=0.57
scipy>=1.12.0
google-generativeai>=0.8.0
